+++ test/strategy.test.js
@@ -31,7 +31,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: undefined,
           customHeaders: undefined
         });
@@ -68,7 +67,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });
@@ -108,7 +106,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });
@@ -146,7 +143,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });
@@ -184,7 +180,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });
@@ -222,7 +217,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });
@@ -260,7 +254,6 @@ describe('Strategy', function() {
           tokenURL: 'https://server.example.com/token',
           userInfoURL: undefined,
           clientID: 's6BhdRkqt3',
-          clientSecret: 'some_secret12345',
           callbackURL: 'https://client.example.org/cb',
           customHeaders: undefined
         });